Divide news card to component
